Refactor pfm-core global
- Convert module into directory.
- Each dependency its own file.
- Use LazyLock, instead lazy_static, for global vars.
- Remove forex_manager, to be implemented later.
- Remove forex_impl APIs tests, test in tests/ instead.
