--- pfm-core/src/forex_manager_impl/mod.rs
@@ -1,1 +1,0 @@
-pub mod forex_manager_storage;
