+++ pfm-core/src/global/constants.rs
@@ -1,0 +1,3 @@
+use crate::forex::Currency;
+
+pub const BASE_CURRENCY: Currency = Currency::USD;
