+++ pfm-core/src/lib.rs
@@ -4,6 +4,3 @@ pub mod forex_impl;
 pub mod global;
 
 pub mod utils;
-
-pub mod forex_manager;
-pub mod forex_manager_impl;
